refactor: reorganize tenant-related client and hook structure
- Replaced TenantClient and TenantMemberClient with new imports from tenant and tenant-member directories.
- Removed the use-tenants hook and its associated functionality to streamline the SDK.
- Added new methods to the SDK class for finding tenants and tenant members by ID.
- Updated type exports to reflect the new structure and removed deprecated tenant types.
