+++ packages/sdk/src/react/hooks/index.ts
@@ -2,4 +2,3 @@ export { useAsyncState } from './use-async-state.js';
 export type { AsyncState, UseAsyncStateReturn } from './use-async-state.js';
 export { useAuth } from './use-auth.js';
 export { useSDK } from './use-sdk.js';
-export { useTenants, useTenantsList } from './use-tenants.js';
