+++ packages/sdk/src/tenant-member/types/paginated-tenant-member-result.type.ts
@@ -1,0 +1,4 @@
+import type { PaginatedResult } from '@/shared/types/index.js';
+import type { TenantMemberResponse } from './tenant-member-response.type.js';
+
+export type PaginatedTenantMemberResult = PaginatedResult<TenantMemberResponse>;
