+++ packages/sdk/src/tenant-member/types/tenant-member-add-input.type.ts
@@ -1,0 +1,8 @@
+import type { TenantMemberRole } from './tenant-member-role.type.js';
+
+export type TenantMemberAddInput = {
+  tenantId: string;
+  userId?: string;
+  role: TenantMemberRole;
+};
+
