+++ packages/sdk/src/tenant-member/types/tenant-member-find-by-criteria-input.type.ts
@@ -1,0 +1,4 @@
+import type { FindByCriteriaInput } from '@/shared/types/index.js';
+
+export type TenantMemberFindByCriteriaInput = FindByCriteriaInput;
+
