+++ packages/sdk/src/tenant-member/types/tenant-member-find-by-id-input.type.ts
@@ -1,0 +1,3 @@
+export type TenantMemberFindByIdInput = {
+  id: string;
+};
