+++ packages/sdk/src/tenant-member/types/tenant-member-remove-input.type.ts
@@ -1,0 +1,4 @@
+export type TenantMemberRemoveInput = {
+  id: string;
+};
+
