+++ packages/sdk/src/tenant-member/types/tenant-member-response.type.ts
@@ -1,0 +1,7 @@
+export type TenantMemberResponse = {
+  id: string;
+  tenantId?: string;
+  userId?: string;
+  role?: string;
+};
+
