+++ packages/sdk/src/tenant-member/types/tenant-member-role.type.ts
@@ -1,0 +1,2 @@
+export type TenantMemberRole = 'OWNER' | 'ADMIN' | 'MEMBER' | 'GUEST';
+
