+++ packages/sdk/src/tenant-member/types/tenant-member-update-input.type.ts
@@ -1,0 +1,7 @@
+import type { TenantMemberRole } from './tenant-member-role.type.js';
+
+export type TenantMemberUpdateInput = {
+  id: string;
+  role?: TenantMemberRole;
+};
+
