+++ packages/sdk/src/tenant/types/paginated-tenant-result.type.ts
@@ -1,0 +1,5 @@
+import type { PaginatedResult } from '@/shared/types/index.js';
+import type { TenantResponse } from './tenant-response.type.js';
+
+export type PaginatedTenantResult = PaginatedResult<TenantResponse>;
+
