+++ packages/sdk/src/tenant/types/tenant-delete-input.type.ts
@@ -1,0 +1,3 @@
+export type TenantDeleteInput = {
+  id: string;
+};
