+++ packages/sdk/src/tenant/types/tenant-find-by-criteria-input.type.ts
@@ -1,0 +1,4 @@
+import type { FindByCriteriaInput } from '@/shared/types/index.js';
+
+export type TenantFindByCriteriaInput = FindByCriteriaInput;
+
