+++ packages/sdk/src/tenant/types/tenant-find-by-id-input.type.ts
@@ -1,0 +1,3 @@
+export type TenantFindByIdInput = {
+  id: string;
+};
